+++ api/rust/include/LIEF/rust/asm/AssemblerConfig.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "LIEF/visibility.h"
+#include <memory>
+#include "LIEF/asm/AssemblerConfig.hpp"
+
+struct AssemblerConfig_r;
+
+LIEF_API std::unique_ptr<LIEF::assembly::AssemblerConfig>
+  from_rust(const AssemblerConfig_r& config);