+++ src/asm/asm.cpp
@@ -81,7 +81,9 @@ Binary::instructions_it Binary::disassemble(const uint8_t*, size_t, uint64_t) co
   return make_empty_iterator<assembly::Instruction>();
 }
 
-std::vector<uint8_t> Binary::assemble(uint64_t/*address*/, const std::string&/*Asm*/) {
+std::vector<uint8_t> Binary::assemble(uint64_t/*address*/, const std::string&/*Asm*/,
+    assembly::AssemblerConfig& /*config*/)
+{
   LIEF_ERR(ASSEMBLY_NOT_SUPPORTED);
   return {};
 }
@@ -269,12 +271,14 @@ Engine::instructions_it Engine::disassemble(const uint8_t*, size_t, uint64_t) {
   return make_empty_iterator<assembly::Instruction>();
 }
 
-std::vector<uint8_t> Engine::assemble(uint64_t/*address*/, const std::string&/*Asm*/) {
+std::vector<uint8_t> Engine::assemble(uint64_t/*address*/, const std::string&/*Asm*/,
+                                      AssemblerConfig& /*config*/)
+{
   return {};
 }
 
 std::vector<uint8_t> Engine::assemble(uint64_t/*address*/, const std::string&/*Asm*/,
-                                      LIEF::Binary&/*bin*/)
+                                      LIEF::Binary&/*bin*/, AssemblerConfig& /*config*/)
 {
   return {};
 }